Add jQuery library and related scripts to documentation
The jQuery library has been added to improve the functionality of documentation. This commit also includes a range of scripts delivered by jQuery, including smooth scrolling for internal links, copying URLs to the clipboard from anchors, and implementing a notification system for clipboard copies. Additionally, styles for the notification system have been included in the stylesheet.